Some big changes, still not done yet. To better fit with the way I want to handle memory, Arrays are now all allocated at once; e.g there is no a->m pointer, just a flexible array member at the end of the struct. The scanner now collects a list of array pointers, rather than arrays, and the scanner's private memory pool stores the entire array struct in a contiguous region. Similarly, the stacks in the evaluators now operate oon this list of arrays, using array pointer pointers for their markers.
It was a big change that required me to touch just about every piece of the system, but at least I made it while the code base was small enough. Array creation/access is now more regular, and controlled through the functions in array.c. There is still a lot of accessing of struct members, but I think that is fine.

I've gotten back to the point where expressions are correctly parsed again. The final roadblock is where I left off; properly storing variables in the henvironment hash table and properly freeing variables with no references.

also added forgotten const.c and memory.c files to repo